+++ src/DefinitionSet/Implementation.php
@@ -1,0 +1,29 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\Math\DefinitionSet;
+
+use Innmind\Math\{
+    Algebra\Number,
+    Exception\OutOfDefinitionSet,
+};
+
+/**
+ * @psalm-immutable
+ * @internal
+ */
+interface Implementation
+{
+    /**
+     * ∈ or ∉
+     */
+    public function contains(Number $number): bool;
+
+    /**
+     * @throws OutOfDefinitionSet
+     */
+    public function accept(Number $number): void;
+    public function union(self $set): self;
+    public function intersect(self $set): self;
+    public function toString(): string;
+}