+++ src/DefinitionSet/Range.php
@@ -10,8 +10,9 @@ use Innmind\Math\{
 
 /**
  * @psalm-immutable
+ * @internal
  */
-final class Range implements Set
+final class Range implements Implementation
 {
     private bool $lowerInclusivity;
     private bool $upperInclusivity;
@@ -93,13 +94,13 @@ final class Range implements Set
     }
 
     #[\Override]
-    public function union(Set $set): Set
+    public function union(Implementation $set): Implementation
     {
         return Union::of($this, $set);
     }
 
     #[\Override]
-    public function intersect(Set $set): Set
+    public function intersect(Implementation $set): Implementation
     {
         return Intersection::of($this, $set);
     }