+++ src/DefinitionSet/RealNumbers.php
@@ -7,8 +7,9 @@ use Innmind\Math\Algebra\Number;
 
 /**
  * @psalm-immutable
+ * @internal
  */
-final class RealNumbers implements Set
+final class RealNumbers implements Implementation
 {
     #[\Override]
     public function contains(Number $number): bool
@@ -23,13 +24,13 @@ final class RealNumbers implements Set
     }
 
     #[\Override]
-    public function union(Set $set): Set
+    public function union(Implementation $set): Implementation
     {
         return Union::of($this, $set);
     }
 
     #[\Override]
-    public function intersect(Set $set): Set
+    public function intersect(Implementation $set): Implementation
     {
         return Intersection::of($this, $set);
     }