+++ tests/DefinitionSet/IntegersExceptZeroTest.php
@@ -4,10 +4,7 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Math\DefinitionSet;
 
 use Innmind\Math\{
-    DefinitionSet\IntegersExceptZero,
     DefinitionSet\Set,
-    DefinitionSet\Union,
-    DefinitionSet\Intersection,
     Algebra\Integer,
     Algebra\Real,
     Exception\OutOfDefinitionSet,
@@ -16,22 +13,14 @@ use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class IntegersExceptZeroTest extends TestCase
 {
-    public function testInterface()
-    {
-        $this->assertInstanceOf(
-            Set::class,
-            new IntegersExceptZero,
-        );
-    }
-
     public function testStringCast()
     {
-        $this->assertSame('ℤ*', (new IntegersExceptZero)->toString());
+        $this->assertSame('ℤ*', Set::integersExceptZero()->toString());
     }
 
     public function testContains()
     {
-        $set = new IntegersExceptZero;
+        $set = Set::integersExceptZero();
 
         $this->assertTrue($set->contains(Integer::of(1)));
         $this->assertTrue($set->contains(Integer::of(-1)));
@@ -41,27 +30,25 @@ class IntegersExceptZeroTest extends TestCase
 
     public function testAccept()
     {
-        $this->assertNull((new IntegersExceptZero)->accept(Integer::of(1)));
+        $this->assertNull(Set::integersExceptZero()->accept(Integer::of(1)));
 
         $this->expectException(OutOfDefinitionSet::class);
         $this->expectExceptionMessage('0 ∉ ℤ*');
 
-        (new IntegersExceptZero)->accept(Integer::of(0));
+        Set::integersExceptZero()->accept(Integer::of(0));
     }
 
     public function testUnion()
     {
-        $union = (new IntegersExceptZero)->union(new IntegersExceptZero);
+        $union = Set::integersExceptZero()->union(Set::integersExceptZero());
 
-        $this->assertInstanceOf(Union::class, $union);
         $this->assertSame('ℤ*∪ℤ*', $union->toString());
     }
 
     public function testIntersect()
     {
-        $intersection = (new IntegersExceptZero)->intersect(new IntegersExceptZero);
+        $intersection = Set::integersExceptZero()->intersect(Set::integersExceptZero());
 
-        $this->assertInstanceOf(Intersection::class, $intersection);
         $this->assertSame('ℤ*∩ℤ*', $intersection->toString());
     }
 }