+++ tests/DefinitionSet/NaturalNumbersTest.php
@@ -4,10 +4,7 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Math\DefinitionSet;
 
 use Innmind\Math\{
-    DefinitionSet\NaturalNumbers,
     DefinitionSet\Set,
-    DefinitionSet\Union,
-    DefinitionSet\Intersection,
     Algebra\Integer,
     Algebra\Real,
     Exception\OutOfDefinitionSet,
@@ -16,22 +13,14 @@ use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class NaturalNumbersTest extends TestCase
 {
-    public function testInterface()
-    {
-        $this->assertInstanceOf(
-            Set::class,
-            new NaturalNumbers,
-        );
-    }
-
     public function testStringCast()
     {
-        $this->assertSame('ℕ', (new NaturalNumbers)->toString());
+        $this->assertSame('ℕ', Set::naturalNumbers()->toString());
     }
 
     public function testContains()
     {
-        $set = new NaturalNumbers;
+        $set = Set::naturalNumbers();
 
         $this->assertTrue($set->contains(Integer::of(1)));
         $this->assertTrue($set->contains(Integer::of(0)));
@@ -41,7 +30,7 @@ class NaturalNumbersTest extends TestCase
 
     public function testAccept()
     {
-        $set = new NaturalNumbers;
+        $set = Set::naturalNumbers();
 
         $this->assertNull($set->accept(Integer::of(1)));
 
@@ -53,17 +42,15 @@ class NaturalNumbersTest extends TestCase
 
     public function testUnion()
     {
-        $union = (new NaturalNumbers)->union(new NaturalNumbers);
+        $union = Set::naturalNumbers()->union(Set::naturalNumbers());
 
-        $this->assertInstanceOf(Union::class, $union);
         $this->assertSame('ℕ∪ℕ', $union->toString());
     }
 
     public function testIntersect()
     {
-        $intersection = (new NaturalNumbers)->intersect(new NaturalNumbers);
+        $intersection = Set::naturalNumbers()->intersect(Set::naturalNumbers());
 
-        $this->assertInstanceOf(Intersection::class, $intersection);
         $this->assertSame('ℕ∩ℕ', $intersection->toString());
     }
 }