+++ tests/DefinitionSet/RealNumbersExceptZeroTest.php
@@ -4,10 +4,7 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Math\DefinitionSet;
 
 use Innmind\Math\{
-    DefinitionSet\RealNumbersExceptZero,
     DefinitionSet\Set,
-    DefinitionSet\Union,
-    DefinitionSet\Intersection,
     Algebra\Integer,
     Algebra\Real,
     Algebra\Value,
@@ -17,22 +14,14 @@ use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class RealNumbersExceptZeroTest extends TestCase
 {
-    public function testInterface()
-    {
-        $this->assertInstanceOf(
-            Set::class,
-            new RealNumbersExceptZero,
-        );
-    }
-
     public function testStringCast()
     {
-        $this->assertSame('ℝ*', (new RealNumbersExceptZero)->toString());
+        $this->assertSame('ℝ*', Set::realNumbersExceptZero()->toString());
     }
 
     public function testContains()
     {
-        $set = new RealNumbersExceptZero;
+        $set = Set::realNumbersExceptZero();
 
         $this->assertTrue($set->contains(Integer::of(1)));
         $this->assertTrue($set->contains(Integer::of(-1)));
@@ -44,7 +33,7 @@ class RealNumbersExceptZeroTest extends TestCase
 
     public function testAccept()
     {
-        $set = new RealNumbersExceptZero;
+        $set = Set::realNumbersExceptZero();
 
         $this->assertNull($set->accept(Integer::of(1)));
 
@@ -56,17 +45,15 @@ class RealNumbersExceptZeroTest extends TestCase
 
     public function testUnion()
     {
-        $union = (new RealNumbersExceptZero)->union(new RealNumbersExceptZero);
+        $union = Set::realNumbersExceptZero()->union(Set::realNumbersExceptZero());
 
-        $this->assertInstanceOf(Union::class, $union);
         $this->assertSame('ℝ*∪ℝ*', $union->toString());
     }
 
     public function testIntersect()
     {
-        $intersection = (new RealNumbersExceptZero)->intersect(new RealNumbersExceptZero);
+        $intersection = Set::realNumbersExceptZero()->intersect(Set::realNumbersExceptZero());
 
-        $this->assertInstanceOf(Intersection::class, $intersection);
         $this->assertSame('ℝ*∩ℝ*', $intersection->toString());
     }
 }