+++ tests/DefinitionSet/RealNumbersTest.php
@@ -4,10 +4,7 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Math\DefinitionSet;
 
 use Innmind\Math\{
-    DefinitionSet\RealNumbers,
     DefinitionSet\Set,
-    DefinitionSet\Union,
-    DefinitionSet\Intersection,
     Algebra\Integer,
     Algebra\Real,
     Algebra\Value,
@@ -16,22 +13,14 @@ use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class RealNumbersTest extends TestCase
 {
-    public function testInterface()
-    {
-        $this->assertInstanceOf(
-            Set::class,
-            new RealNumbers,
-        );
-    }
-
     public function testStringCast()
     {
-        $this->assertSame('ℝ', (new RealNumbers)->toString());
+        $this->assertSame('ℝ', Set::realNumbers()->toString());
     }
 
     public function testContains()
     {
-        $set = new RealNumbers;
+        $set = Set::realNumbers();
 
         $this->assertTrue($set->contains(Integer::of(1)));
         $this->assertTrue($set->contains(Integer::of(0)));
@@ -43,7 +32,7 @@ class RealNumbersTest extends TestCase
 
     public function testAccept()
     {
-        $set = new RealNumbers;
+        $set = Set::realNumbers();
 
         $this->assertNull($set->accept(Integer::of(1)));
         $this->assertNull($set->accept(Integer::of(0)));
@@ -55,17 +44,15 @@ class RealNumbersTest extends TestCase
 
     public function testUnion()
     {
-        $union = (new RealNumbers)->union(new RealNumbers);
+        $union = Set::realNumbers()->union(Set::realNumbers());
 
-        $this->assertInstanceOf(Union::class, $union);
         $this->assertSame('ℝ∪ℝ', $union->toString());
     }
 
     public function testIntersect()
     {
-        $intersection = (new RealNumbers)->intersect(new RealNumbers);
+        $intersection = Set::realNumbers()->intersect(Set::realNumbers());
 
-        $this->assertInstanceOf(Intersection::class, $intersection);
         $this->assertSame('ℝ∩ℝ', $intersection->toString());
     }
 }