+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Requires PHP `8.4`
+- `Innmind\Math\DefinitionSet\Set` is now a final class, all previous implementations are now flagged as internal
 
 ### Fixed